backend/requirements.txt
aiohttp==3.9.3
python-dotenv==1.0.1
fastapi==0.110.0
pydantic==2.6.3
pytest==8.0.2
starlette==0.36.3
uvicorn==0.27.1
requests==2.31.0
httpx==0.22.0